new module for the manager
