+++ limbus/core/__init__.py
@@ -1,5 +1,6 @@
-from .core import ComponentsManager, Component, DefaultParam, ComponentState, Params, NoValue
-from .factory import ComponentDefinition, register_components
+from limbus.core.core import Component, ComponentState, Params, NoValue
+from limbus.core.manager import ComponentsManager, DefaultParam
+from limbus.core.factory import ComponentDefinition, register_components
 
 __all__ = [
     "ComponentsManager",
